Fix layout logic in IrArray::Index::Linearize.
The logic previously assumed the layout was major-to-minor. In practice, I believe this assumption is correct in all places where Linearize is called, because LayoutNormalization converts most instructions to be major-to-minor. But in general, IrArray supports arbitrary layouts and so we should obey the given layout in IrArray::Index::Linearize as well. I am working on a change which calls IrArray::Index::Linearize in another place, and it requires Linearize to correctly handle non-major-to-minor layouts.

Also fix a comment incorrectly stating multidim was major-to-minor.

PiperOrigin-RevId: 566971940