Make `PjRtMemorySpace::id()` unique within a client
The current API semantics requires memory space ids to be unique only within a memory kind. This is inconsistent with `PjRtDevice::id()` (unique within a client regardless of device types) and cumbersome (since `(memory_kind, id)` is needed to uniquely identify a memory space). This CL changes the contract to require implementations to guarantee ids to be unique within a client.

PiperOrigin-RevId: 567117355