1) Fix: * test_dense_input_ragged_weights_fails, * test_sparse_input_ragged_weights_fails.
sparse_ops.sparse_bincount is intended to handle inputs of dense tensors (including non-TensorFlow types that can be converted to dense tensors) and sparse tensors only. Ragged tensor inputs are expected to use other software via dispatch. `weights` are not considered by dispatch. Previously, if `weights` is not a sparse tensor, it is assumed to be something that can be converted to a dense tensor (e.g. a numpy array or Python list) but this conversion fails for ragged tensors. This CL only converts `weights` if it is not any type of composite tensor, allowing error checking to detect any mismatch in the type of tensor between the input and `weights`.

2) Enable running ragged bincount tests.

PiperOrigin-RevId: 567132514