Disable two compiler warnings in the bazelrc
This is adding `-Wno-error=unused-command-line-argument` and `-Wno-gnu-offsetof-extensions` to the `rbe_linux_cpu` config.

We have these flags already in the `release_cpu_linux` config,
but some RBE builds don't use that config. So I'm adding
them to `rbe_linux_cpu` as well.

Ideally there would be a config that applies to all clang builds but not to the
old GCC builds, but that doesn't exist. So for now I don't see a good option
other than having the flag in two places (without major restructurings of the bazelrc file).

PiperOrigin-RevId: 567330212