Make SimplifyFPConversions check creation pass ids
This is adding an additional model to `SimplifyFPConversions` which will only
simplify chains of converts if they were generated by a previous optimization pass.

The pass is checking whether these convert ops were created by an optimization pass
or part of the input HLO by checking the `logical_creation_pass_id` metadata field.
A pass id of -1 means they were part of the input.

Note that this changing is only preparing the pass and adding changes, but it is
not yet changing any behaviour in XLA.

PiperOrigin-RevId: 568467656