[xla] Modify P2PSchedulePreparation to work with the CollectivPermuteDecomposer and handle pipelined Send-Recv chains.
The P2PSchedulePreparation pass now amends control dependence to express the
intended ordering of an unpipelined Send-Recv chain, as the
CollectivePermuteDecomposer no longer generates such control dependence.

Add control dependence to ensure that in a whild-body with pipelined a Send-Recv
chain, Send the ordered before Recv.

Add control dependence to ensure if an instruction with nested Send-Recv and
also used the result of a while-loop with pipelined Send-Recv, the instruction
is ordered after the pipelined send-done.

Add tests.

PiperOrigin-RevId: 568899565