Fix race condition when two concurrent function executions use the same resource input.
Previously, there was a time-of-check-to-time-of-use race in `EagerRemoteExecute()`:

1. First, we check `input->HasResourceShapeAndMirror()`. If the result is `false`, we'll set `serialize_resource_dtype_and_shape` to `true`.

2. If `serialize_resource_dtype_and_shape` is `true`, the call to `SerializeRemoteTensorHandle()` will populate `input_handle->resource_dtypes_and_shapes()`.

3. If `input_handle->resource_dtypes_and_shapes()` is populated, we'll call `input->AddResourceShapeMirror()`.

These steps were not atomic, so if two threads execute step (1) concurrently and observe no mirror, they'll both try to add one in (3), and the second attempt will fail.

This change makes `TensorHandle::AddResourceShapeMirror()` idempotent. In the event of concurrent access, we will do some extra work on the first attempt, but this avoids more complicated synchronization.

PiperOrigin-RevId: 570723881