For aarch64, use empty kTargetCpuForHost string
The string "aarch64" does not appear in the list of valid CPUs in
AArch64.td, making it not a valid processor in
llvm/lib/MC/MCSubtargetInfo.cpp:getFeatures, which reports an
"unrecognized CPU" error, ultimately hitting an llvm_unreachable
assertion during code lowering when running tests such as
cpu_eigen_dot_operation_test. Log of erroneous behavior below.

This is very similar to the x86_64 machine description, which also
does not have a true "x86_64" processor listed. Using the
empty string as the cpu type gets translated to "generic", which is
handled properly by the rest of the infra.

Match that behavior for aarch64, which allows the test to pass properly.

Failing log for cpu_eigen_dot_operation test when run on an aarch64 host:
"""
[==========] Running 6 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 6 tests from CpuEigenDotOperationTestInstantiation/CpuEigenDotOperationTest
[ RUN      ] CpuEigenDotOperationTestInstantiation/CpuEigenDotOperationTest.SimpleDotOp/F16
'aarch64' is not a recognized processor for this target (ignoring processor)
'aarch64' is not a recognized processor for this target (ignoring processor)
'aarch64' is not a recognized processor for this target (ignoring processor)
'aarch64' is not a recognized processor for this target (ignoring processor)
'aarch64' is not a recognized processor for this target (ignoring processor)
'aarch64' is not a recognized processor for this target (ignoring processor)
Don't know how to custom expand this
UNREACHABLE executed at llvm/lib/Target/AArch64/AArch64ISelLowering.cpp:24052!
"""

PiperOrigin-RevId: 571415910