[hlo_evaluator] Don't create blank, zero-populated Literals for tuples
When evaluating Tuple(), create a blank Literal for the destination that is populated with
kUndetermined.

We already copy in all known leaves, so all this does is avoid us creating blank leaf Pieces,
which is incredibly important when dealing with large programs pre-sharding. Also note that
we'd instantaneously throw the blank pieces all away...

PiperOrigin-RevId: 571532410