Add int4 support to IrArray.
In LLVM, an int4 IrArray is represented as an array of i8 values, where each i8 value stores two packed 4-bit integers. The reason an array of i4 values is not used is because i4 values are not packed in LLVM IR (the size of any type is a multiple of 1 byte).

Also add tests that check the generated LLVM IR for both the int4 and non-int4 cases.

PiperOrigin-RevId: 572138947