Lift done callback out of cuda context scope
This change fortifies the codebase against random deadlocks.
The TF Executor may inline another kernel in the successful done callback,
which expects the original cuda context and hangs if the finishing kernel does not
restore the context.

Note that some of the use of the cuda context scope appears to be wider than
necessary. This CL does not attempt to narrow done the scopes due to uncertainty
 of the level of test coverage and author's inexperience in this topic.

PiperOrigin-RevId: 572649699