Don't allow to fuse into the first operand of scatter.
This allows to emit the scatter in-place. We also need
to avoid forming fusions where using the in-place emitter
would run into race conditions. Also, adjust the buffer
sharing logic so that it also detects for scatter fusions
whether the buffer of a fusion operand can be shared with
the output.

PiperOrigin-RevId: 572869882