#tf-data-service Snapshot manager recovers failed splits.
It fixes this error:
```
Found missing global split index, 1565, in /path/to/dataset_snapshot
```

If the dispatcher fails when writing a split, the split
should be recovered. Otherwise, the dispatcher cannot find
the split with the specified global index.

**Fix:**
Instead of skipping a split, rewrite the split to the split file.

PiperOrigin-RevId: 573048774