[PJRT C API] Change CheckMatchingStructSizes to be ActualStructSizeIsGreaterOrEqual.
With this change, it is no longer required that the plugin has the same struct size as the framework. Because ActualStructSizeIsGreaterOrEqual is called in the plugin, this means it will check whether framework has a greater or equal struct size than the plugin (meaning the framework is newer than the plugin).

PiperOrigin-RevId: 573059095