Fix analysis of partial epilogue-fused reductions.
When the epilogue crosses a fusion instruction, we currently do
not analyze this correctly, because we are unable to follow edges
through fusion instructions (they are considered non-intermediate).
The fix is to reuse the logic from HloTraversal for these edges:
when encountering a parameter or fusion, continue outside or inside
the fusion, respectively.

PiperOrigin-RevId: 573151739