Priority fusion: remove heuristics, model performance instead.
Currently, priority fusion applies some heuristics to prevent bad
fusion decisions (e.g. fusing a reduce into a transpose). These
heuristics are hard to debug and follow, so it is better to cost
model the effect instead.

For now, ee do this with a rather crude coalescing analysis. We
assume memory accesses are coalesced iff there is a) no
transpose or b) we can use the transpose emitter.

Drive-by fix: use the proper launch grid for DUS. We always
assume it's in place. We're doing this the wrong place, but it's the only
place we have for now.

The new `ShouldFuse` function was contributed by kramerb@.

PiperOrigin-RevId: 573215172