PR #6318: [NVIDIA XLA] Use a seperate mutex for callback_stream_map to avoid deadlocks during shutdown.
Imported from GitHub PR https://github.com/openxla/xla/pull/6318

In `LocalDeviceState`, currently one mutex `mu_` is used to guard multiple things. In  `ThenExecuteCallback()` it is used to guard `callback_stream_map_` while in `ReturnStreamToPool()` it is used to guard `usage_stream_pool_`.

I have encountered a deadlock during executor shutdown where the mutex is being held by `ThenExecuteCallback()` while it is waiting on `cuEventRecord`. At the same, a cuda callback which runs `ReturnStreamToPool()` is blocked while it waits for the mutex. The stalled cuda callback prevents forward progress on the stream and cuEventRecord is unable to finish.

To fix this, I added a separate mutex for `ThenExecuteCallback` which is only used to guard `callback_stream_map_`, to avoid the false dependency on changes to `usage_stream_pool_`.
Copybara import of the project:

--
1b52f38384376d0764c81b61336779aa3d31b5a3 by Trevor Morris <tmorris@nvidia.com>:

Use a seperate mutex for callback_stream_map to avoid deadlocks during shutdown.

--
0641594fffff8c2c815f3ad77012c4c175abe801 by Trevor Morris <tmorris@nvidia.com>:

Fix build warnings

Merging this change closes #6318

PiperOrigin-RevId: 573504792