[XLA:GPU] Don't test tsl::enable_tensor_float_32_execution(false) for dot anymore
The `dot` instruction doesn't support the flag directly, just by checking the operand precisions.
tf2xla/transforms/legalize_tf.cc sets them to highest if the TensorFloat-32 global variable is false.
So now we also do that in the test.
Hopefully b/280130359 will handle this in a nicer way.

This came up when trying to enable xla_gpu_triton_gemm_any. When we do that, we'll switch a lot of dots over to the Triton codegen, which ignores the TensorFloat-32 global variable and only checks the operand precisions.

PiperOrigin-RevId: 574452125