#tf-data-service Fix dataset file already exists error.
A previous fix [1] did not address all the cases. If the user
has a requested a dataset ID, this could also happen. So apply
the fix to both cases. Since the dispatcher state is the source
of truth, if the state does not have a dataset [2], any existing
dataset file should be overwritten.

[1] https://github.com/tensorflow/tensorflow/commit/25c61d2ca860ad7f0a842f2f9052dfe67dca9391
[2] https://github.com/tensorflow/tensorflow/blob/33c0f436e79d671e4fbd0aa63b449f1fb1a9be12/tensorflow/core/data/service/dispatcher_impl.cc#L586

PiperOrigin-RevId: 575358928