Define `CustomAggregatorIdAssigner` class to utilize python implementation from the cpp layer.
Newly defines `CustomAggregatorIdAssigner` class and exposes it to the python layer through pybind11.
The new class has a pure virtual function `AssignIds` which is exposed as `assign_ids` to the python side for overriding.
In the python layer, the method is overridden with a python implementation, which can then be called and utilized from the cpp layer.

PiperOrigin-RevId: 576331382