Add SM 9.0 / Remove SM 7.5 from TF builds
With the recent CUDA update we gained support for SM 9.0 compute capabilities (Hopper-based GPUs),
but so far we didn't enabled this by default in the nightly builds.

To keep file sizes in check we remove the build for SM 7.5. User's of the
Turing architecture (SM 7.5) will fall back to Volta (SM 7.0) which is compatible.

The main difference between Turing and Volta are TensorCores which TF doesn't target from custom CUDA kernels, hence we don't expect any peformance disadvantages
for Turing users.

It also brings us in line with what JAX is doing.

PiperOrigin-RevId: 576835080