[XlaCallModule] Improved support for effects and version 9.
Previously, XlaCallModule op supported only at most one token and it
had to appear at the start of the arguments and results. Now we look
for token types among the input and result types. This allows us to
support more than one token and not necessarily at the start. This is
useful when the module is serialized from JAX, which uses one token
for each distinct effect, and the tokens may appear after the platform
index argument in presence of multi-platform lowering.

We also fix the bitcast operations introduced prior to shape refinement
to add them only when needed: for non-tokens with dynamic shapes.

This introduces version 9 of XlaCallModuleOp.
See https://github.com/google/jax/blob/main/jax/experimental/jax2tf/README.md#native-serialization-versions
for a description of the different versions.

PiperOrigin-RevId: 577032011