[tsl] forward TSL logging to Absl logging
Remove the LOG primitives in "default" and just rely on Absl.
These primitives were created when binary size was a concern.
That concern is not an issue anymore, so remove the primitives
and rely on Absl's logging.

With this change code can now depend on Absl directly, except
when needing some parts of Absl that aren't yet in OSS, namely VLOG.

PiperOrigin-RevId: 577935032