Fix a test that tries to use a non-assertion method as an assertion.
This test is currently broken in two ways:
1. Usage of `called_once_with` as an assertion. All that does is create
that method on the mock (exposed in Python3.12, which warns about this).
The right thing to use would have been `assert_called_once_with`.

2. Usage of the same function for testing twice, which doesn't work, as it's
cached inside an allowlist the first time, and won't error the second time.
This was hidden by the first issue.

PiperOrigin-RevId: 578915953