Fix a bug in the `xla-call-module-deserialization` pass where function name uniquification didn't consider name conflicts correctly
The current `NewFuncName()` implementation identifies a new function name that is unique in the target module to which StableHLO functions are inserted, but it does not check whether the newly generated name clashes with any existing function in the original StableHLO module. For example, if there are two functions `@foo` and `@foo0` in the StableHLO module and the target TF module has `@foo`, `NewFuncName()` will rename the first function `@foo` into `@foo0` as this name does not exist in the TF module, but this causes name collision between the two StableHLO functions.

This CL rewrites the renaming logic to leverage `mlir::SymbolTable::renameToUnique()`. This allows us to rename func names to be unique to both modules without having to manually generate func names, which simplifies the code.

PiperOrigin-RevId: 578922049