Fix issue with element_size_on_bits on int4 arrays in XLA:CPU.
LayoutAssignment might remove the element_size_in_bits layout field, so now SubByteNormalization will add it back. We already did this on GPUs, but on CPUs, for some reason, the element_size_in_bits fields is only removed in certain cases, such as when the array is multidimensional, so I didn't notice this had to be done on CPUs before.

PiperOrigin-RevId: 578930077