Create `PyFunctionLibrary` to specify functions overriding cpp virtual functions.
Moved `CustomAggregatorIdAssigner::AssignIds` to `PyFunctionalLibrary`. This new class is expected to be used as a collection of python-overridden functions.
This change also factors out the custom `type_caster`s to a separate file at `type_casters.h`, because these are both used in `pywrap_function_lib.cc` and `pywrap_quantize_model.cc`.

PiperOrigin-RevId: 579350552