[XLA:SPMD] Add a missing matching case for collective matmul (windowed einsum).
When we may group on both lhs and rhs, if we can compute the number of iterations for only a single one, we should pick that one as the matching. The unit test is extracted from a bug, where the number of iterations was only computed for lhs, but lhs_matching was set to false because of the missing case.

PiperOrigin-RevId: 579886356