Avoid C++14/17 features in TFLM shared code
The flatbuffer_conversions.cc source is shared between TFLite and TFLite-Micro via a nightly sync. While TFLite's C++ standard is C++17, TFLite-Micro must remain on C++11 due to use with older toolchains. This change removes the usage  of a few C++14/17 features so that flatbuffer_conversions.cc can be used in TFLite-Micro:

* Use of auto in lambda parameters (C++14) - This was used to handle varying parameter array sizes in the `LoadAttr` lambda. It is replaced with an `int64_t*` and size (in bytes). This also means that the compiler won't generate multiple versions of the `LoadAttr` lambda, one for each array size.
* Use of init-statements within if statements (C++17) - This was used for checking the resulting TfLiteStatus. It is replaced with using `TF_LITE_ENSURE_STATUS`.

PiperOrigin-RevId: 580118011