Remove unnecessary "self" from test_util.py decorators.
In cases where the `self` keyword is unused in a decorator function
and its presence is not necessary for other reasons (e.g. passing
positional arguments after `self` in the decorated function), removes
`self` from the signature of the returned function and from the invocation
of the wrapped/decorated function.

This change also modifies the returned function(s) to be one returned by
`tf_decorator.make_decorator` to align with other decorators in the file
and to ensure necessary metadata (such as the original function's argspec)
is preserved.

In all of these cases, the calls should be functionally equivalent.

PiperOrigin-RevId: 580237930