Install libtpu-nightly from an index
LibTPU is necessary for Cloud TPU VMs. It is no longer included within the TensorFlow TPU wheels. Instead, it is downloaded from an index. When pip install TensorFlow TPU wheels, add the flag `-f <libtpu_download_url>` to the command to download the libtpu wheel.

libtpu.so has been compiled with a toolchain that is too new for manylinux2014. Since it is no longer packaged within the TF-TPU wheels, the TF-TPU wheels should be tagged with manylinux2014 rather than manylinux_2_27.

Remove unused TFCI LIBTPU env vars.

PiperOrigin-RevId: 580719053