Fix misleading variable name in SetBufferHandleToInput test: s/kInputTensorIndex/kOutputTensorIndex/
This was probably a cut-and-paste error from SetBufferHandleToOutput below,
which uses the same variable name for the tensor at index 3.
In the graph constructed by TestDelegation (which TestDelegate inherits from),
which is used by both of these tests, the tensors at index 0 and 1 are inputs
and tensors at index 3 and 4 are outputs.

PiperOrigin-RevId: 580730411