PR #5782: [NVIDIA XLA GPU] Turn the cudnn fused attention feature ON by default
Imported from GitHub PR https://github.com/openxla/xla/pull/5782

Now that CUDA 12 and cuDNN 8.9.4 are now in CI and the feature is in a steady state, it is time to turn this ON by default and reap the benefits.
Note that this doesn't include Flash Attention.
Copybara import of the project:

--
309d03a4b1fc8573b161118cd43361652c3880f4 by Ayan Moitra <amoitra@nvidia.com>:

Turning the cudnn fused attention feature on by default in XLA:GPU

--
1ea48af1f3f8007559348c7dbb1f86734c8d49bb by Ayan Moitra <amoitra@nvidia.com>:

Fix test failures

Merging this change closes #5782

PiperOrigin-RevId: 581190924