Move run_without_tensor_float_32 decorators to the correct location.
PR #59140 incorrectly applied the decorator to the outer function instead
of applying it to the interior, test function. Since the decorator's internal
function does not return a value, this modified the test template functions
to return `None` instead of their internal test functions.

When these returned `None` objects were then passed to
`test_util.run_deprecated_v1`, that function returned a decorator.
This decorator was then added to the test class under the test name.

In other words, the application of this decorator to the exterior function instead
of the interior had two effects:
- It disabled the actual test from running since the decorated version of the
  function returned `None`.
- It replaced the tests with invocations of the decorator returned by
  `test_util.run_deprecated_v1`.

This change simply relocates the decorator to the interior function, undoing
the above changes while preserving the original intent of the PR.

This change also disables a very small set of tests which fail upon being
re-enabled by this CL. Since they were not running before this change, this is
not a regression introduced by this CL.

PiperOrigin-RevId: 581359387