Do not mangle async error in various pjrt buffer implementations.
Before this change, at various points in pjrt buffer implementations, when the
async buffer becomes available as an absl::Status error, the error code is
replaced by absl::StatusCode::kInternal, and the error message is also
modified.

This change propagates the underlying error without modification to the
callback waiting on that async value.

PiperOrigin-RevId: 581987519