Updated the description of max_delta parameter
The documentation currently states that the max_delta parameter controls the maximum absolute change in brightness. However, the max_delta parameter actually controls the maximum relative change in brightness. This means that the actual change in brightness will depend on the range of values in the input image. So, I have updated the description for the max_delta parameter. Could you please have a look and do the needful. Thank you!