Multi-threaded fusion priority evaluation.
This improves the overall time we spend on HLO passes by ~14% (priority
vs. priority). The other low-hanging fruit is caching of analysis results.
We rerun it a lot right now: in each priority computation, we do it for the
producer, for each consumer, and for each producer-consumer fusion. The
first two can be cached. Similarly, in GpuPriorityFusion::ChooseKind. This
is still a todo.

PiperOrigin-RevId: 582670292