Update test_util decorators to be either decorators or decorator factories (not both).
This change modifies most of the decorator functions in `test_util.py`
to make them behave consistently as either decorators or as decorator
factories.

That is, the modified functions no longer conditionally return a decorator
or a decorated function depending on if the passed-in function is `None`.

If the decorator accepted optional parameters other than the function, it was
updated to be a decorator factory (a function which returns a decorator).
If it did not have any additional parameters, it was updated to only work
as a decorator.

For example:
```
def foo(config=True, func=None):
  def decorator(f):
    def decorated(*args, **kwargs):
      ...
      return f(*args, **kwargs)
    return decorated
  if func is not None:
    return decorator(func)
  return decorator
```
would now be:
```
def foo(config=True):
  def decorator(f):
    def decorated(*args, **kwargs):
      ...
      return f(*args, **kwargs)
    return decorated
  return decorator
```

while:
```
def bar(func=None):
  def decorator(f):
    def decorated(*args, **kwargs):
      ...
      return f(*args, **kwargs)
    return decorated
  if func is not None:
    return decorator(func)
  return decorator
```
became:
```
def bar(func):
  def decorated(*args, **kwargs):
    ...
    return func(*args, **kwargs)
  return decorated
```

This change also modifies internal TensorFlow code which used these
decorators, if necessary, in order to comply with the changes.

PiperOrigin-RevId: 583036777