Add type annotations to stack.py and traceable_stack.py.
Changes `DefaultStack` into a parameterized class which allows
users to specify the type of object stored on that stack. Also adds
generic parameters to `TraceableStack` and `TraceableObject` to allow
the same type-checking specialization (if desired).

Notably does not change the session stack to use this new generic type
as that would introduce a cyclic dependency between `stack.py` and `client/session.py`.

PiperOrigin-RevId: 583388633