PR #7088: Fix incorrect device memory limit on non root processes.
Imported from GitHub PR https://github.com/openxla/xla/pull/7088

@hawkinsp

In a multiprocess config, this code was attempting to get the memory limit from device 0 which is only local to process 0. For all other processes, the memory limit was being set to 0.

This was causing the memory limit in the latency hiding scheduler to be different between process 0 and the other processes, leading to a different scheduled order of collectives and causing deadlocks.

To fix this, use addressable_devices instead of all global device and index using the device_ordinal.
Currently, jax doesn't appear to set `options.executable_build_options.device_ordinal()` so we will default to 0 in that case.

Copybara import of the project:

--
1bad102e374c7aa4c3acfefdad66477813432a82 by Trevor Morris <tmorris@nvidia.com>:

Use addressable devices so that correct mem limit is read. Also use device_ordinal to index into addressable devices - however device_ordinal is currently unset so will default to 0

Merging this change closes #7088

PiperOrigin-RevId: 583928913