Fix misuse of ArrayRef
Fix some uses of ArrayRef that result in out of scope accesses
being made. This includes accesses on the stack after being freed.
Also fixed a concat attempt of two arrays through ArrayRef that
was producing nonsense.
The issues were located with the address sanitizer and shown to
be fixed with these changes.