Add type annotations to tf_contextlib.
Adds type annotations to `tf_contextlib.contextmanager` to allow type
checkers to infer that functions annotated with this decorator now return
a context manager - specifically one which yields the same value as the
function being wrapped.

This should improve type checking when context managers are
manually handled and invoked outside of `with` statements.

PiperOrigin-RevId: 584353646