[XLA:GPU] Fuse the scopes of the dot fusion to separate subgraphs
I tried to enable Triton fusions for some new cases, and I found that some of the fusions contained dead code (unreachable instructions).
The reason was that some instructions were reachable both through the LHS and the RHS.
The BFS (Breadth-first search) through the LHS operand "marked" one operation as non-fusible because it would exceed the limit on fusion parameters per operand.
But the BFS through the RHS operand went through that node and fused some more operands.
So the resulting fusion was not connected and caused errors.
This CL fixes the issue by generating separate subgraphs for each scope.

The different scopes are loaded separately anyway, so this change shouldn't slow down anything.

This doesn't change the softmax fusion.

Sorry for the complicated test case.

PiperOrigin-RevId: 584567910