PR #7269: Fix an incorrect static_assert
Imported from GitHub PR https://github.com/openxla/xla/pull/7269

The size of a uint8_t is 1 so a static_assert
to check that it is 0 makes no sense, fix it.
Also fix a couple of warnings about lack of
typename
Copybara import of the project:

--
e72164f0a16ec5e548927ca3583d1a31edbd95d8 by Andrew Goodbody <andrew.goodbody@linaro.org>:

Fix an incorrect static_assert

The size of a uint8_t is 1 so a static_assert
to check that it is 0 will always be false.
The original intent was to have the assert only
trigger if the struct was instantiated but the
standard deems it ill formed if it can never be
true and allows compilers to reject it. Adopt a
different workaround that avoids this by allowing
the possibility of an evaluation to true.
Also fix a couple of warnings about lack of
typename

Merging this change closes #7269

PiperOrigin-RevId: 585767871