Add type annotations to test_util.py.
Adds parameter and return type annotations to the majority of public functions
and methods in the `test_util` module. This includes annotations for methods on
`TensorFlowTestCase` which return values, but omits the assertion methods.

If adding types is currently infeasible
(due to complexity of the signature,
limitations of the supported versions of python, type checker limitations, etc.),
then this change simply does not add those annotations.

PiperOrigin-RevId: 586008562