[XLA:GPU] Add a pass merging producer fusions in to Triton Softmax fusions.
This merges producer fusions into triton softmax fusions provided that the resulting fusion would be able to be tiled. The producer fusions can have multiple inputs, but for now must have a single output (the triton softmax fusion). The merger does not yet operate on consumer fusions or with any basic cost modeling, both of which will be added in a followup CL.

PiperOrigin-RevId: 586193270