validate axis argument of tf.squeeze
In current implementation of tf.squeeze, if we pass a tensor for axis argument it will raise error like below, which is not inferable.

MemoryError: std::bad_alloc

This is due to the reason that the Squeeze Op expects axis as a list and there is no check for this in C++ level. Hence I am adding a validation check at Python level itself.