Add overload of `ModifyGraphWithDelegate` that takes a `TfLiteOpaqueDelegate*`.
This is similar to the existing overload of `InterpreterBuilder::AddDelegate`.
When building with TFLITE_USE_OPAQUE_DELEGATE, this overload is needed to use ModifyGraphWithDelegate with delegates created using TfLiteOpaqueDelegateCreate.

InterpreterBuilder::AddDelegate remains preferred over Interpreter::ModifyGraphWithDelegate, but this new overload is useful
because it reduces the burden for converting code to use the opaque delegate APIs.
PiperOrigin-RevId: 586663060