Remove tensorrt dependency when installing tensorflow[and-cuda]
tensorrt's odd packaging is causing install headaches (TF 2.15.0
currently cannot be installed because it depends on tensorrt_ packages).
Also, even if it is installed correctly, TF doesn't seem to be able to use it
(it says: could not find tensorrt). For the time being, our solution
is to remove this dependency for the simple [and-cuda] install method.

Note: tensorrt is packaged differently than other Nvidia packages.
The "tensorrt" package, when installed, fetches two other packages
from Nvidia's PyPI repository that are not present on the main
PyPI repository.
PiperOrigin-RevId: 587089327