Update configure.py
This edit is made to address the bug in the configuration where the build system ignores the PYTHON_BIN_PATH environment variable, which was one of the issues raised in the last few days.

I followed the style and code of conduct proposed for making changes in the Tensorflow library. Also, I have added comments that explain what the code is doing.

I hope you will take these edits in consideration, have a great day!