Improve the speed of TF Quantizer
The majority of time spent on the quantizer is to run CustomAggregatorOp. This cl limits the CustomAggregatorOp insertion to before and after composite functions. This help speeding up the conversion. For resnet v2, conversion time is reduced from 30-80 minutes to 10-20 minutes.

For gather ops, CustomAggregatorOp is only inserted before them since it is weight only quantization. This cl also enforces min_num_elements_for_weights for gather ops in the SRQ path.

PiperOrigin-RevId: 588605491