[XLA:GPU] Add Consumer Fusion merging to FusionMergerTriton
FusionMergerTriton now tries to merge consumer fusions into triton softmax fusions. The consumer will try to merge irrespective of is the producer merge is successful. The consumer may take additional operands, which (if fused) will also be operands of the resulting merged fusion. The consumer may not be multi output at this time.

PiperOrigin-RevId: 588889684