[jax] Make from_dlpack error on non-default layouts.
We're currently working on better layout support in jax, but for now,
jax can't really handle Arrays with non-standard layouts. Raise an
error instead of returning wrong results when the Array is passed to a
computation (!). This can be reverted once jax has better support for
non-default layouts.

Related issues:
https://github.com/google/jax/issues/7657
https://github.com/google/jax/issues/17784

PiperOrigin-RevId: 589258555