Simplify local reproduction instructions
This change is a big simplification of reproducing a build locally:

- Moved a lot of configurable behavior into source-able local_* env files
- Deleted a lot of extraneous instructions
- Added "how do I choose..." instructions for clarification
- Fixed a bug preventing multiple invocations of rename_and_verify_wheels
- Made any.sh and bisect.sh use the cleaner method of sourcing local_ envs
- Cleaned up ci_default and added "how to see an overview of env variables"

One of the big things is that I added some logic in local_default that resets
the list of Bazel common args to remove things like --config=rbe. This way a
user can choose to run a "release" env configuration without being blocked by
permission errors, or by needing to manually amend their configuration. As a
result, gathering the correct set of config values is quite easy (see the new
instructions). The alternative would be to split "auth-related" flags into
a separate variable, which I don't like, as I think it's more confusing for
future maintainers to have to continuously decide which options go where.