[XLA:GPU] Move ShouldFuse logic inside the priority queue.
Right now, ShouldFuse is called in the queue and in InstructionFusion::Run. There is no reason to do the latter, because fusion queue fully controls what should be fused.

Duplication logic from InstructionFusion::ShouldFuse is not applicable for PriorityFusion. We only need ShouldFuseInPlaceOp which is a static function already.

Saves us a tiny bit of compile time, because we don't recompute the logic on the slow single-threaded pass.

PiperOrigin-RevId: 589781890