Validate mask argument of tf.boolean_mask
As per tf.boolean_mask documentation the argument mask must be:
K-D boolean Tensor, K <= N and K must be known statically.

Currently the API tf.boolean_mask not validating the mask argument. If we pass any numeric valued tensors/lists irrespective of positives or negatives all will be considered as True except for 0. This may not produce desirable results. IMO, it's better to enforce the user to pass the mask as boolean tensor. Hence proposing the validation.