Redo the batch size calculation in HorizontalLoopFusion.
The final horizonal loop fusion should not become too big. Right now, it is
limited by the number of fusion instructions we allow to horizontally fuse.
This does not account for the number of fusion outputs in multi-output fusions,
which also affects the final size of the horizontal loop fusion. So now we
count the number of fusion outputs instead of just the number of fusion
instructions. This makes the code generation much faster, and we can re-enable
the test that recently started to time out after changes in LLVM.

PiperOrigin-RevId: 590083251