#tf-data-service Fix a deadlock when loading a distributed snapshot dataset.
The original `FirstComeFirstServedTaskRunner::GetProcessingTimeNsec`
acquires a mutex lock that guards `iterator_`. In load v2,
`iterator_->GetNext` may be blocked when waiting for the snapshot to
complete. However, the dispatcher won't be able to know the snapshot
is complete without worker heartbeats, which calls `GetProcessingTimeNsec`.
This results in a deadlock.

Fix: The FirstComeFirstServedTaskRunner can expose the Model object,
which has its own locks. Calling `model->ComputeSnapshotProcessingTimeNsec`
does not block the iterator lock.

An additional benefit of this fix is that the next time we need something
from Model, we don't need to repeat all the delegations.

PiperOrigin-RevId: 590290283