[xla] Disable decomposition for collective-permute without a channel id.
This is because a program can have multiple collective-permute without a channel
id. If we decompose such collective-permute and use channel id 0 for the
created send and recv, the existing approach in p2p-schedule-preparation pass
can't reconstruct the send and recv chains.

Change the tests to use explicit channel ids. Add a test for the no channel id
case.

PiperOrigin-RevId: 591002454