#tf-data-service Add repetitions to `test_distributed_load` test.
Also fix a bug in repeat dataset op: if we reset the split provider
first, there may still be threads that are calling `GetNext` of the
split provider. For example, the CallFunctions of the parallel map
dataset. This may cause the split provider to exhaust the next
repetition and increment its repetition counts. The next time the
iterator gets the next split, it could see end_of_splits due to
mismatching repetition counts.

PiperOrigin-RevId: 591324122