Add conversion from HloSharding replicated to ShardingParam.
This change renames sharding_param_to_op_shading to shading_conversions as the conversions can be in either direction, and adds a minimal conversion from replicated HloSharding to ShardingParam. The change adds only this minimal case in order to be able to test the upcoming IFRT IR Python bindings. In the future, we'll be adding support for conversions between more types of HloShardings and ShardingParams.

PiperOrigin-RevId: 592286258