[XLA:GPU][TileAnalysis] Add experimental tile abstraction and initial tests.
A tile describes a structured subset of indices inside a N-dimensional array,
where the set of indices captured along each dimension can be expressed as a
strided expression
```
    offset + stride * iota(size)
```
with offset, stride, and size three non-negative integers, and iota the usual
range function.

A N-dimensional symbolic tile is a function from offsets, strides, and sizes
to a N-dimensional tile. It is encoded as an affine map
```
    (stride0, offset0, ..., stride{M-1}, offset{M-1})[size0, ... size{P-1}]
 -> (expr0, ..., expr{N-1})
```
where `expr0, ... expr{N-1}` are strided expressions as described above.

Symbolic tiles also store, for each one of their parameters, what its
upper bound is (accessible through `max_sizes()` for size parameters and
`max_strides_and_offsets()` for offset and stride parameters). Size
parameters may also be assigned a specific value (accessible through
`sizes()`).

Symbolic tiles are constructed from the shape of the N-dimensional array we
want to tile, or by propagating (composing) an existing tile with an
`IndexingMap`. Tile propagation may fail if the results of the produced
affine map are not all strided expressions.

PiperOrigin-RevId: 595337323