[XLA:GPU] Roll forward of generalized triton tiling propagation of concatenations
It had originally rolled back due to numerical issues with concat that have since been fixed.

Store offsets along the concatenated dimensions in the dimension orders of operands to reflect that they may be incompatible with other uses of the same operands. This allows fusing concatenations which are not the only users of their inputs.

PiperOrigin-RevId: 595458030