[XLA] Add ConditionalCallInstruction to HloComputation.
Follows the pattern of the recently added `WhileCallInstruction` and `CollectiveCallInstruction`.

There is existing logic that a computation should not be reassigned to an instruction of a different type. I didn't find a good explanation for the reason of it, except for fusions. I had to restructure hlo_instruction_test that was reusing the same computation for `conditional` and `while`.

This is getting out of hand. I tried to convert all those back pointer instructions into one member, but it turned out to be a much bigger refactoring than I expected.

PiperOrigin-RevId: 596554457