[XLA:GPU] Triton: Use arith.minimumf/maximumf on Ampere+
NaN propagation used to be broken on these ops but it was fixed in
https://github.com/openai/triton/commit/6efcc926d63e2663ad76ed7e9036345033ace584

This compiles down to min.NaN in ptx, which is a single instruction on sm_80+
and thus a bit faster for compute bound kernels. Semantics are covered by
existing tests.

Sadly the old code has to stay around for Volta as LLVM cannot lower minimumf
there. It also requires passing down the compute capability way down into the
emitter.

PiperOrigin-RevId: 596586000