Fix race condition in kernel caching.
If two identical kernels are created at the same time, they will both observe that the cache doesn't contain their cache_key, and both try to create the kernel at the same time. The first kernel to be created gets added to the cache, then the second kernel overwrites the first kernel in the cache.

This can cause issues for stateful kernels such as kernels that own resources. Destroying the first kernel in the above example could cause in-use resources to be destroyed.

This CL resolves the issue by checking if a cache entry already exists right before adding to the cache. If we create a new kernel but discover a cached entry has been added during the kernel creation, we now discard the new kernel and use the cached entry. This way we only ever use one kernel per cache key.

PiperOrigin-RevId: 596762121