Add requests Python library to dependencies in Python package
TF used to be implicilty depend on requests for quite some time
through `python/distribute/failure_handling/failure_handling_util.py`,
and `requests` got installed automatically because it was a dependency
of `tb-nightly`.

But since `tb-nightly` does not depend on `requests` anymore we now get
`Module not found` errors. The solution is to explicitly depend on
`requests`.

PiperOrigin-RevId: 596957313