Use correct `group_dim_sizes` in `GetGroupedReplicatedSharding`.
In `GroupedSharding`, `group_dim_sizes` is the number of groups along a specific dimension. For example, if `group_dims[i] = j, group_dim_sizes[i] = k`, it means that we form k groups along the j-th dimension. Thus, `Product(group_dim_sizes)` is the total number of groups, which should be equal to `device_groups.size()`.

In `GetGroupedReplicatedSharding`, `group_dim_sizes` should be `{num_groups}`, instead of `{group_size}`. We also add (1) a CHECK in UngroupSharding, and (2) a test case, accordingly.

This cl is a preparation of cl/595212770. More details: go/refactor-hlo-sharding-util

PiperOrigin-RevId: 597079931