Remove the appearance of the word `mlir` in DebugOptions.
The plan is to split the `DebugOptions` proto target into header-only and implementation targets to prevent ODR violations.
However, unfortunately all symbols that contain the word `mlir` is not exported from libtensorflow_framework.so because it is blocked in the version script (`unexported_symbols_list` for MacOS).
To avoid the member fields of `DebugOptions` from being blocked, this change tries to remove the appearance of `mlir` from symbols.

PiperOrigin-RevId: 597101951