Simplify parallel_loop_emitter index computation code a bit.
The logic isn't as different for the first element and the rest, or for
row_vectorized/not row vectorized as it looks.

Should be an NFC (and the tests seem to agree). I'm not sure why the
multidim optimization is only enabled for the row_vectorized case.
Maybe some secret invariant about thread_idx_x that's not documented here?
Also not sure if this optimization does anything in the first place, but
that's a question for a different day.

PiperOrigin-RevId: 597171848