Clean up wrong callers of FindNonTrivialHero.
We had several callers of FindNonTrivialHero which called it with a fusion
instruction. This was usually an indicator that the call was not actually needed
at all and we should just use the instruction itself as the hero.
While there, also fix an inconsistency in IsUniversallyLoopFusible which did only
check for copy ops whether they can be emitted with the transpose emitter. But
it looks like this will make no difference, as we will still allow fusing via the
IsInputFusible() check.
Also adjust ChooseFusionKind to check the producer, too, whether it is a kInput
fusion.

PiperOrigin-RevId: 597209606