[XLA:GPU] Reuse nodes when fusing operands if the iterspec is the same
This avoids code duplication during the compilation when it's not needed and makes the reuse of the values explicit.

The fusion graph will more closely resemble the final executable, in the sense that nodes are duplicated when we need 2 different reads and they are not duplicated when one read is enough.

The reuse now happens at the level of FuseTowardOperands, so we don't reuse nodes if they are used in different "operand trees" (LHS, RHS, operands of the nodes fused to the output of the dot).

Now we represent the fusion (while it's being built) by a traditional graph representation (FusionGraph), instead of GraphPath which I introduced earlier, and remove now.

PiperOrigin-RevId: 597258953