This is change 2/2 for allowing the user to tell the HeapSimulator how to iterate over slice time permutations. (A permutation of slice times tells the compiler when to start each slice allocation.) In the future, we want MSA to be able to trim the iteration space for the HeapSimulator based on higher-level knowledge.
In this change we make the SlicedAllocationFinder explicitly take a SliceTimePermutationIterator, so the user can specify which permutations to try. In addition, we move the code to check for legal permutations from the best fit repacker to the HeapSimulator, so all iterators can use it.

This is mostly a refactoring.

PiperOrigin-RevId: 597340523