[XlaCallModule] Fixes for serialization version 9.
In version 9, the main function of a serialized module may contain token arguments and outputs. Those do not correspond to actual XlaCallModule op inputs and outputs.

In cl/577032011 we had adjusted the input_shapes for the call to RefineDynamicShapes from xla_call_module_op. Here we move the adjustment to
input_shapes inside the RefineDynamicShapes, so that it takes effect for all
call sites, including those from shape_inference.

PiperOrigin-RevId: 597445704