Fix heap-use-after-free in GpuCompiler
At the beginning of `OptimizeHloModule` we get a reference to the `debug_options` inside of
the `config` property of the `HloModule`. The config property is a copy-on-write structure
which means our reference to the debug options becomes dangling as soon as someone
changes the config property. This happens for example when the `SubByteNormalization` pass changes the entry computation layout from the config.

To avoid the issue I have replaced the local reference variable by inlining the call to `hlo_module->config().debug_options()`. There was already a mixture between using the reference and going through hlo_module which is now consistent.

PiperOrigin-RevId: 597580503