Add a counter to track replicated graphs without a device type
Currently MlirBridgePass checks both the device type of a graph and if it's replicated to decide which phase 1 pipeline to run. If we know a replicated graph always contains a device type, then we can remove the check on device types. The counter added will be used to verify that there are no replicated graphs without device type in production jobs.

PiperOrigin-RevId: 597916417