Ensure the allocation type is kTfLiteCustom when doing shallow copies in DeepOrShallowCopyTensorsShapeTypeData.
This code is correct only under the assumption that the caller has correctly prepared
the tensors that get passed in for shallow copying, by setting their allocation
types to kTfLiteCustom. This ensures that those tensors won't be double `free`'d
later on. This check simply ensures that that assumption always holds, to ensure
we fail early if ever a bug is introduced that breaks that assumption.

PiperOrigin-RevId: 597990125