[XlaCallModule] Fix multi-platform support.
Previously, the XlaCallModuleLoader::Create would load the module and preprocess it to set the platform index argument for modules that were serialized for multi-platform support.

We remove the setting of the platform index from the `Create` function and we create a new method `SetPlatformIndex` that is called when the op is compiled, before shape refinement. This does not change the semantics in most use cases, when we have a sequence of `Create` and `Compile`.

The only semantic difference from previous usage should be when the op is pre-processed, e.g., for shape refinement, in which case we call `Create` without then calling `Compile`. And the semantic difference is only in presence of multi-platform modules, in which case we used to get an error.

PiperOrigin-RevId: 598097856